Better organization of player lists (#230)
