+++ client/src/GamemasterInGame/PlayerLists/NightPlayerList.tsx
@@ -5,107 +5,23 @@ import {
   Flex,
   Heading,
   IconButton,
-  Separator,
   Text,
 } from "@radix-ui/themes";
-import { PlayerNameWithRoleIcon, RoleName } from "../shared/RoleIcon";
+import { PlayerNameWithRoleIcon } from "../../shared/RoleIcon";
 import { getCharacter } from "@hidden-identity/shared";
-import { GiAxeInStump, GiBootKick } from "react-icons/gi";
 import { RxHamburgerMenu } from "react-icons/rx";
-import React, { HTMLAttributes, useState } from "react";
-import { useDefiniteGame } from "../store/GameContext";
-import { DeadVoteIcon, PlayerStatusIcons } from "./NotesIcons";
-import { PlayerList } from "./PlayerListComponents";
-import { DemonMessage } from "./PlayerListComponents/PlayerMessage/DemonMessage";
-import { DialogHeader } from "../shared/DialogHeader";
-import {
-  PlayerFilter,
-  PlayerListFilters,
-  usePlayerFilters,
-} from "../shared/PlayerListFilters";
-import { useKickPlayer } from "../store/actions/playerActions";
-import { DestructiveButton } from "./DestructiveButton";
+import React from "react";
+import { useDefiniteGame } from "../../store/GameContext";
+import { PlayerList } from "../PlayerListComponents";
+import { DemonMessage } from "../PlayerListComponents/PlayerMessage/DemonMessage";
+import { DialogHeader } from "../../shared/DialogHeader";
+import { DestructiveButton } from "../DestructiveButton";
 import { Role } from "@hidden-identity/shared";
-import {
-  PlayerListOrder,
-  PlayerOrder,
-  usePlayerOrder,
-} from "../shared/PlayerListOrder";
-import {
-  usePlayerNotes,
-  useVotesToExecute,
-} from "../store/actions/gmPlayerActions";
-import { Reveal } from "../types/PlayerMessageScreen";
-import { FaFeather } from "react-icons/fa6";
-import { useFirstSeat, useIsHiddenView } from "../store/url";
-import { SetCountModal } from "../shared/SetCount";
+import { usePlayerNotes } from "../../store/actions/gmPlayerActions";
+import { Reveal } from "../../types/PlayerMessageScreen";
+import { useIsHiddenView } from "../../store/url";
+import { PlayerNotes } from "../PlayerListComponents/PlayerNotes";
 
-export function PregamePlayerList() {
-  const { game } = useDefiniteGame();
-  const [, kickPlayerLoading, , handleKickPlayer] = useKickPlayer();
-  const seatingProblems =
-    game.orderedPlayers.problems && game.orderedPlayers.playerProblems;
-
-  return (
-    <Flex className="overflow-y-auto" direction="column" py="3" gap="2">
-      {Object.entries(game.playersToRoles).length === 0 && (
-        <Text as="div" className="m-5 text-center">
-          No players have joined yet. Share the game by clicking the game code.
-        </Text>
-      )}
-      {Object.entries(game.playersToRoles).map(([player, role]) => (
-        <Flex
-          justify="between"
-          align="center"
-          px="3"
-          gap="3"
-          key={player}
-          asChild
-        >
-          <Text size="2">
-            <PlayerList.Name player={player} />
-            {seatingProblems ? (
-              <Text as="div">
-                {seatingProblems[player] ? "Getting settled" : "Ready"}
-              </Text>
-            ) : (
-              <div
-                className="truncate capitalize"
-                style={{
-                  flex: 2,
-                }}
-              >
-                {RoleName(role)}
-              </div>
-            )}
-
-            <Dialog.Root>
-              <Dialog.Trigger>
-                <IconButton variant="ghost">
-                  <RxHamburgerMenu />
-                </IconButton>
-              </Dialog.Trigger>
-              <Dialog.Content className="m-2">
-                <Flex direction="column" gap="2">
-                  <Dialog.Close>
-                    <PlayerList.MenuItem
-                      id="kick-player"
-                      label="Kick Player"
-                      onClick={() => handleKickPlayer(player)}
-                      disabled={kickPlayerLoading}
-                    >
-                      <GiBootKick />
-                    </PlayerList.MenuItem>
-                  </Dialog.Close>
-                </Flex>
-              </Dialog.Content>
-            </Dialog.Root>
-          </Text>
-        </Flex>
-      ))}
-    </Flex>
-  );
-}
 type Action =
   | {
       type: "character";
@@ -335,102 +251,5 @@ export function NightPlayerList({
         </DestructiveButton>
       )}
     </Flex>
-  );
-}
-
-interface IngamePlayerListProps {
-  selectedOrder: PlayerOrder;
-  setSelectedOrder: (order: PlayerOrder) => void;
-}
-export function IngamePlayerList({
-  selectedOrder,
-  setSelectedOrder,
-}: IngamePlayerListProps) {
-  const { game } = useDefiniteGame();
-  const [firstSeat] = useFirstSeat();
-  const orderedPlayers = usePlayerOrder(selectedOrder, firstSeat);
-  const allFilters = usePlayerFilters(orderedPlayers);
-  const [selectedFilter, setSelectedFilter] = useState<PlayerFilter>("all");
-  const filteredPlayers = allFilters[selectedFilter];
-  const [, , , setVotesToExecute] = useVotesToExecute();
-
-  return (
-    <Flex className="h-full overflow-y-auto" direction="column" p="2" gap="2">
-      <PlayerListFilters
-        allFilters={allFilters}
-        selectedFilter={selectedFilter}
-        setSelectedFilter={setSelectedFilter}
-      />
-      <PlayerListOrder
-        selectedOrder={selectedOrder}
-        setSelectedOrder={setSelectedOrder}
-      />
-      <Separator size="4" />
-      {filteredPlayers.map((player) => (
-        <Flex direction="column" key={player}>
-          <Text size="4" asChild>
-            <Flex justify="between" align="center" gap="3">
-              <PlayerList.RoleIcon player={player}>
-                {getCharacter(game.playersToRoles[player]).ability}
-              </PlayerList.RoleIcon>
-              <PlayerList.NoteInputModal player={player}>
-                <button className="flex-1 text-left">
-                  <PlayerList.Name player={player} />
-                </button>
-              </PlayerList.NoteInputModal>
-              <DeadVoteIcon player={player} />
-              <SetCountModal
-                title="Set votes to execute:"
-                onSet={(votes: number) => setVotesToExecute(player, votes)}
-              >
-                <IconButton variant="soft" radius="large">
-                  <GiAxeInStump />
-                </IconButton>
-              </SetCountModal>
-              <PlayerList.Actions player={player}>
-                <IconButton id={`${player}-menu-btn`} variant="ghost">
-                  <RxHamburgerMenu />
-                </IconButton>
-              </PlayerList.Actions>
-            </Flex>
-          </Text>
-
-          <PlayerNotes className="px-[3em] py-1" player={player} />
-        </Flex>
-      ))}
-    </Flex>
-  );
-}
-
-interface PlayerNotesProps extends HTMLAttributes<HTMLDivElement> {
-  player: string;
-}
-
-function PlayerNotes({ player, ...props }: PlayerNotesProps) {
-  const { game } = useDefiniteGame();
-  const statuses = game.playerPlayerStatuses[player] ?? [];
-  const notes = game.playerNotes[player] ?? "";
-  if (!notes && !statuses.length) return;
-
-  return (
-    <Text size="2" weight="light" asChild>
-      <Flex direction="column" gap="2" {...props}>
-        {statuses.length > 0 && (
-          <Flex gap="3">
-            <PlayerStatusIcons player={player} />
-          </Flex>
-        )}
-        {notes && (
-          <PlayerList.NoteInputModal player={player}>
-            <button className="ml-1 flex-1 whitespace-pre-line text-left">
-              <Flex gap="2">
-                <FaFeather />
-                {notes}
-              </Flex>
-            </button>
-          </PlayerList.NoteInputModal>
-        )}
-      </Flex>
-    </Text>
   );
 }
