+++ client/src/GamemasterInGame/PlayerLists/SetupPlayerList.tsx
@@ -1,0 +1,74 @@
+import { Dialog, Flex, IconButton, Text } from "@radix-ui/themes";
+import { useDefiniteGame } from "../../store/GameContext";
+import { useKickPlayer } from "../../store/actions/playerActions";
+import { PlayerList } from "../PlayerListComponents";
+import { RxHamburgerMenu } from "react-icons/rx";
+import { GiBootKick } from "react-icons/gi";
+import { RoleName } from "../../shared/RoleIcon";
+
+export function SetupPlayerList() {
+  const { game } = useDefiniteGame();
+  const [, kickPlayerLoading, , handleKickPlayer] = useKickPlayer();
+  const seatingProblems =
+    game.orderedPlayers.problems && game.orderedPlayers.playerProblems;
+
+  return (
+    <Flex className="overflow-y-auto" direction="column" py="3" gap="2">
+      {Object.entries(game.playersToRoles).length === 0 && (
+        <Text as="div" className="m-5 text-center">
+          No players have joined yet. Share the game by clicking the game code.
+        </Text>
+      )}
+      {Object.entries(game.playersToRoles).map(([player, role]) => (
+        <Flex
+          justify="between"
+          align="center"
+          px="3"
+          gap="3"
+          key={player}
+          asChild
+        >
+          <Text size="2">
+            <PlayerList.Name player={player} />
+            {seatingProblems ? (
+              <Text as="div">
+                {seatingProblems[player] ? "Getting settled" : "Ready"}
+              </Text>
+            ) : (
+              <div
+                className="truncate capitalize"
+                style={{
+                  flex: 2,
+                }}
+              >
+                {RoleName(role)}
+              </div>
+            )}
+
+            <Dialog.Root>
+              <Dialog.Trigger>
+                <IconButton variant="ghost">
+                  <RxHamburgerMenu />
+                </IconButton>
+              </Dialog.Trigger>
+              <Dialog.Content className="m-2">
+                <Flex direction="column" gap="2">
+                  <Dialog.Close>
+                    <PlayerList.MenuItem
+                      id="kick-player"
+                      label="Kick Player"
+                      onClick={() => handleKickPlayer(player)}
+                      disabled={kickPlayerLoading}
+                    >
+                      <GiBootKick />
+                    </PlayerList.MenuItem>
+                  </Dialog.Close>
+                </Flex>
+              </Dialog.Content>
+            </Dialog.Root>
+          </Text>
+        </Flex>
+      ))}
+    </Flex>
+  );
+}
